Full v9 read and write

+++ src/juicebox/Context.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2011-2017 Broad Institute, Aiden Lab
+ * Copyright (c) 2011-2020 Broad Institute, Aiden Lab
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,8 +24,8 @@
 
 package juicebox;
 
+import juicebox.data.basics.Chromosome;
 import juicebox.windowui.HiCZoom;
-import org.broad.igv.feature.Chromosome;
 
 /**
  * @author jrobinso
@@ -62,9 +62,9 @@ public class Context {
         this.zoom = zoom;
     }
 
-    public int getChrLength() {
-        return chromosome.getLength();
-    }
+    public long getChrLength() {
+		return chromosome.getLength();
+	}
 
     public Chromosome getChromosome() {
         return chromosome;
